Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,6 +4,6 @@ int main() {
     int x,y,z,avg;
     scanf("%d""%d""%d", &x,&y,&z);
     avg = (x + y + z)/3;
-    printf("Average: %.2f\n", avg);
+    printf("Average: %.2f\n", (float)avg);
     return 0;
 }
